requirements.txt
asgiref==3.6.0
Celery==5.2.7
daphne==4.0.0
Django==4.1.5
django-debug-toolbar==4.0.0
django-environ==0.10.0
django-extensions==3.2.1
django-filter==23.1
django-cors-headers==3.14.0
django-import-export==3.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
gunicorn==20.1.0
gevent==22.10.2
geventhttpclient==2.0.9
psycopg2-binary==2.9.6
pytz==2023.3
redis==4.5.4
sentry-sdk==1.19.1
dj-database-url==1.3.0
pyinstrument==4.4.0
Pillow==9.5.0
channels==4.0.0
channels-redis==4.1.0
whitenoise==6.4.0
google-api-python-client==2.85.0
google-auth-httplib2
google-auth-oauthlib
